102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-long long f[50];
+unsigned long f[50];
 int n;
 
 f[0] = 1;
@@ -20,9 +20,9 @@ f[n] = f[n - 1] + f[n - 2];
 
 for (n = 0; n < 49; n++)
 {
-printf("%lld, ", f[n]);
+printf("%lu, ", f[n]);
 }
-printf("%lld\n", f[49]);
+printf("%lu\n", f[49]);
 
 return (0);
 }
